# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Messenger API",
        "description": "1) Получи токен: POST /api/auth/login или /api/auth/register\n2) В Swagger UI нажми Authorize (замок справа) и введи: Bearer <token>\n3) После этого все защищённые запросы будут отправляться с заголовком Authorization.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Базовый URL API"
        }
    ],
    "paths": {
...
        "/api/chats/{chat}/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Список сообщений чата",
                "operationId": "ca139938433ad42e57c2f93997d7862f",
                "parameters": [
                    {
                        "name": "chat",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество на странице (по умолчанию 10)",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        },
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список сообщений",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Пагинация Laravel (data/links/meta)",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Message"
                                            }
                                        },
                                        "links": {
                                            "nullable": true
                                        },
                                        "meta": {
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Отправить сообщение в чат",
                "operationId": "e56581288f5ce635142a2345ff0e35cc",
                "parameters": [
                    {
                        "name": "chat",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Привет!"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "body": "Привет!"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Сообщение отправлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "$ref": "#/components/schemas/Message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "422": {
                        "description": "Ошибки валидации"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/chats/{chat}/messages/search": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Поиск сообщений в чате",
                "operationId": "d2b8210fb7aa0bc7722e0eb48b11aca0",
                "parameters": [
                    {
                        "name": "chat",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Привет"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1
                        },
                        "example": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результаты поиска",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Message"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "422": {
                        "description": "Ошибки валидации"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/{message}": {
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "Удалить сообщение только для себя",
                "operationId": "afa7ae5fe58a4a6dc87f66b85e2e7211",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 501
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удалено"
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/{message}/all": {
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "Удалить сообщение для всех",
                "operationId": "3bb26f7ad673af96fdb831ee268676ba",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 501
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удалено"
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Нет прав"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/{message}/forward": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Переслать сообщение в другой чат",
                "operationId": "be3d30bdcd2b834a5fb59f5e8c56faf2",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 501
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "chat_id"
                                ],
                                "properties": {
                                    "chat_id": {
                                        "type": "integer",
                                        "example": 55
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "chat_id": 55
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Сообщение переслано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "$ref": "#/components/schemas/Message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "422": {
                        "description": "Ошибки валидации"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "description": "Пользователь",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "nickname": {
                        "type": "string",
                        "example": "johnny"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe",
                        "nullable": true
                    },
                    "avatar_path": {
                        "type": "string",
                        "example": "/storage/avatars/1.png",
                        "nullable": true
                    },
                    "avatar_thumb_path": {
                        "type": "string",
                        "example": "/storage/avatars/thumbs/1.png",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Message": {
                "description": "Сообщение",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 501
                    },
                    "body": {
                        "type": "string",
                        "example": "Привет!"
                    },
                    "sender_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "chat_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-02-04T12:30:00Z"
                    },
                    "sender": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Получите токен через /api/auth/login или /api/auth/register и передавайте в заголовке: Authorization: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}